backend/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.14
django-cors-headers==3.12.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.2.0
geographiclib==1.52
geopy==2.2.0
idna==3.3
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
pycodestyle==2.8.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
